Commonize duplicated getXPropertyValues/extractValue IT test helpers
OEClientModelSettingsIT.getModelPropertyValues/extractValue and
OEClientTaskSettingsIT.getTaskPropertyValues/extractValue were
byte-for-byte identical except for the resource URL used (model URI
vs task graph URI). Moved the shared logic into a single
AbstractModelScopedIT.getPropertyValues(sysGraphUri, propertyUri)
method (both IT classes already extend AbstractModelScopedIT), and
reduced the two call sites to thin one-line wrappers.
